RES-3463: inline const eval helper validation
